Added my contribution about the Game Design Document (GDD).
These notes should be considered a "Design Opinion"; they are not final.
A votation is needed, in which all members voice their opinions and feelings about the Music, Art, some Mechanics and maybe other decisions.
After listening to everyone's opinion, the final decision is already made... so the work should be split between the teammates (this should be done, according to the Milestones set (to use our time wisely): on Sunday 21/05/2023 ).  Then this very day ( Sunday ) all members should be able to start working on their own Systems & Game Mechanics.