I am done with the experiment of the Jammo (The Robot) Character Controller, for today (there's more testing to do, tomorrow).
The good news: is that I made it work with Cinemachine, in the Level 1 Prototype.
What is missing:  I will add the "reverse Left vs. Right Buttons feature" tomorrow.
(Also: the performance is decreasing since I added 2 Jammo Characters... we will have to look into some performance Optimization with the URP Batcher, Static Batching for the Scene, Atlas for the Textures (I can do that, there is a tool for that...), etc).

I will go to sleep quickly in a moment, after commiting my changes to the main (also, after merging).
@everyone Please let me know if the merging result in the remote "main" (origin branch) is also successful and working in your PCs, after pulling it from the server, tomorrow morning (if you can). If there are any issues, I will fix them quickly.